Move shared attr filters logic to attribute module
